+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 # Development
  
+* Add support for Podcasting 2.0 `<podcast:remoteItem>` tag (channel level)
 * Add support for Podcasting 2.0 `<podcast:season>` tag
 
 # 1.0.6